rework example to add clarity between supabase client and server side
